requirements.txt
beautifulsoup4==4.9.3
dj-database-url==0.5.0
Django==3.1.7
django-extensions==3.0.8
django-js-asset==1.2.2
djangorestframework==3.11.1
Flask==1.1.2
Flask-Cors==3.0.10
gunicorn==20.1.0
psycopg2-binary==2.8.6
pytz==2021.1
requests==2.9.1
six==1.15.0
whitenoise==2.0.6
